add and configure packit
this commit add configuration for the packit service.
this will enable automatic copr builds for PRs and releases.

it also enables some downstream automation in koji and bodhi which
will trigger downstream changes in dist git upon commits there.
upstream builds to koji are submitted in the form of scratch builds.

Signed-off-by: Stephen Coady <scoady@redhat.com>